feat(rel): support AE 1.7.99 addresslib format 5
Skyrim SE/AE 1.7.99 introduces address-library format 5: a dense
uint32_t[] array direct-indexed by id, replacing the old sparse
delta-encoded format 2 on-disk. Add runtime format dispatch so a
single build keeps loading every format it has ever supported,
selected from the format byte actually in the file rather than a
compile-time toggle.

Split IDDatabase into IDDB/Offset2ID (their own files, mirroring
shad0wshayd3's upstream rewrite), while keeping this fork's lazy-init
+ atomic-flag lifecycle and test-injection/reset hooks instead of a
bare REX::Singleton, since Module::clear() and the test suite need
to reset/reinject address-library state between cases.

Format 5's body is mapped directly from the file (no unpack step),
requiring a new file-backed memory_map::create_from_file() and a
CreateFileW wrapper in REX::W32 (previously only anonymous shared
sections were supported).

Bump SKSE::RUNTIME_SSE_LATEST_AE to 1.7.99.
